Sonic ERaZor - The Final Legacy [Codename]
"Who knows, maybe I will continue this game at some point again. So, unlike last year, where I said Version 3 would be the final release but changed my mind, I will say right away that chances are that I will continue this thing after some time has passed." - Selbi (Revision 86 - July 24th 2012)

Well, I said I would be back and so I am! The past year was enough time for me to list some of the issues and bugs still remaining. On top of that I have a bunch of new ideas I want to get into this thing. Whether it's going to be an update in the size of the fourth public release can only be answered by the future, but I don't really think so.

This first revision is mostly to backup some of the stuff I have been working on in private already.

- [!] 200% more explosions: Each primary explosion spawns two children, and they all move in random directions.
- [!] Atmospheric Mode has been replaced with Nonstop Inhuman Mode, and it unlocks after finishing the game.
- [!] Camera Shaking system has been drastically improved (taken from Sonic MBV).
- Overworld entrances have been covered with a platform to avoid accidentally falling into them (press Down to enter).
- Story and tutorial texts have been edited (thanks to Kaidan and Abcron (both are now in the credits)).
- Easter Egg story has been decensored and put back into the original shape.
- Experimental sinewave gradient background for the options menu (still needs a lot of fix-up).
- Fixed Buzz Bombers playing explosion sounds faster than they should.
- Fixed Peelout sound occasionally looping forever (by making it fade out like the spindash sound).

git-svn-id: http://svn.hapisan.com/erazor@89 27e8f02d-d7b1-461d-98b4-6824791911a2